+++ coil-video/src/main/java/coil/request/Videos.kt
@@ -1,4 +1,3 @@
-@file:Suppress("unused")
 @file:JvmName("Videos")
 
 package coil.request