+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [2.0.0-rc03] - April 11, 2022
+
+- Remove the `ScaleResolver` interface.
+- Convert `Size` constructors to functions.
+- Change `Dimension.Pixels`'s `toString` to only be its pixel value.
+- Guard against a rare crash in `SystemCallbacks.onTrimMemory`.
+- Update Coroutines to 1.6.1.
+
+## [2.0.0-rc02] - March 20, 2022
+
+- Revert `ImageRequest`'s default size to be the size of the current display instead of `Size.ORIGINAL`.
+- Fix `DiskCache.Builder` being marked as experimental. Only `DiskCache`'s methods are experimental.
+- Fix case where loading an image into an `ImageView` with one dimension as `WRAP_CONTENT` would load the image at its original size instead of fitting it into the bounded dimension.
+- Remove component functions from `MemoryCache.Key`, `MemoryCache.Value`, and `Parameters.Entry`.
+
 ## [2.0.0-rc01] - March 2, 2022
 
 Significant changes since `1.4.0`:
@@ -14,9 +29,10 @@ Significant changes since `1.4.0`:
     - You **should not** use OkHttp's `Cache` with Coil 2.0 as the cache can be corrupted if a thread is interrupted while writing to it.
     - `Cache-Control` and other cache headers are still supported - except `Vary` headers, as the cache only checks that the URLs match. Additionally, only responses with a response code in the range [200..300) are cached.
     - Existing disk caches will be cleared when upgrading to 2.0.
-- `ImageRequest`'s default `Scale` is now `Scale.FIT`
+- `ImageRequest`'s default `Scale` is now `Scale.FIT`.
     - This was changed to make `ImageRequest.scale` consistent with other classes that have a default `Scale`.
     - Requests with an `ImageViewTarget` still have their `Scale` auto-detected.
+- `ImageRequest`'s default size is now `Size.ORIGINAL`.
 - Rework the image pipeline classes:
     - `Mapper`, `Fetcher`, and `Decoder` have been refactored to be more flexible.
     - `Fetcher.key` has been replaced with a new `Keyer` interface. `Keyer` creates the cache key from the input data.